+++ src/Api/Technology/technology.repository.ts
@@ -1,12 +1,18 @@
 import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
 import { Model } from "mongoose";
+import { EmployeeRepository } from "../Employee/employee.repository";
+import { ProjectRepository } from "../Project/project.repository";
 import { CreateTechnologyDto, UpdateTechnologyDto } from "./dto/technology.dto";
 import { Technology, TechnologyDocument } from "./technology.schema";
 
 @Injectable()
 export class TechnologyRepository {
-    constructor(@InjectModel(Technology.name) private technologyModel: Model<TechnologyDocument>) {}
+    constructor(
+        @InjectModel(Technology.name) private technologyModel: Model<TechnologyDocument>,
+        private projectRepo: ProjectRepository,
+        private employeeRepo: EmployeeRepository    
+    ) {}
     
     async create(newTechnology: CreateTechnologyDto) {
         const newTechnologyProject = new this.technologyModel(newTechnology)
@@ -45,6 +51,16 @@ export class TechnologyRepository {
     }
 
     async delete(_id: string) {
-        return this.technologyModel.findByIdAndDelete(_id)
+        const findTech = await this.technologyModel.find({_id: _id})
+        if (findTech){
+            const techInProject = await this.projectRepo.findOne({technology:_id})
+            const techInEmployee = await this.employeeRepo.findOne({technology:_id})
+            if ((!techInProject || techInProject.length == 0) && (!techInEmployee || techInEmployee.length == 0)) {
+                await this.technologyModel.findByIdAndDelete(_id)
+                return "You have successfully deleted"
+            }
+            else throw new HttpException("You can not delete", HttpStatus.NOT_ACCEPTABLE)
+        }
+        else throw new HttpException("Not found",HttpStatus.NOT_FOUND)
     }
 }
